update to compile with ogre v.1.8
- change 'size_t' to 'unsigned int' in renderwindow_bind.cpp
- change 'ms_Singleton' to 'msSingleton' in root_bind.cpp

+++ interface/renderwindow_bind.cpp
@@ -48,7 +48,7 @@ void render_window_set_visible(RenderWindowHandle window_handle, int visible)
     window->setVisible(visible);
 }
 
-size_t render_window_get_hwnd(RenderWindowHandle window_handle)
+unsigned int render_window_get_hwnd(RenderWindowHandle window_handle)
 {
 	size_t windowHnd = 0;
 	Ogre::RenderWindow* window = reinterpret_cast<Ogre::RenderWindow*>(window_handle);
